requirements.txt
# Core Dependencies
python-telegram-bot==21.0.1
kiteconnect==5.0.1
openai==1.12.0
streamlit==1.31.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
celery==5.3.4

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Processing
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0
scikit-learn==1.4.0

# Async Support
asyncio==3.4.3
aiohttp==3.9.3
asyncpg==0.29.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
loguru==0.7.2

# Configuration
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Visualization (for Streamlit)
plotly==5.18.0
altair==5.2.0
matplotlib==3.8.2

# ML for Anomaly Detection
isolation-forest==0.0.2
statsmodels==0.14.1

# Time Series
ta==0.11.0  # Technical Analysis library
yfinance==0.2.33  # For historical data in backtesting

# API Rate Limiting
ratelimit==2.2.1
backoff==2.2.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
croniter==2.0.1